requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.8
langchain-community==0.3.8
langchain-core==0.3.21
langchain-ollama==0.2.0
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
langgraph==0.2.53
langgraph-checkpoint==2.0.6
langgraph-sdk==0.1.36
langsmith==0.1.146
marshmallow==3.23.1
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.4.1
openai==1.55.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
propcache==0.2.0
pydantic==2.10.2
pydantic-core==2.27.1
pydantic-settings==2.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.35
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
yarl==1.18.0
